Adding workFactorMillis as config property to receiver sleep time
By adding a longer [configurable] timeout for the message processing
it is easier to simulate a failure during message processing by simply
quitting the application during the thread sleep time. A sleep time of
greater than 5000 milliseconds will more or less guarantee that any
message being processed will not finish during the soft shutdown grace
period, which makes for a more obvious test of the manual-ack.